requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.12.1